Factor out the bottom bar and layer it under the side menu (close #243)
